Move gathering openstack data to async celery task
